Implemented configurable filtering and sorting.

+++ RXL.Core/Server.cs
@@ -9,7 +9,7 @@ namespace RXL.Core
         public uint Players { get; set; }
         public uint Bots { get; set; }
         public uint MaxPlayers { get; set; }
-        public long Latency { get; set; }
+        public uint Latency { get; set; }
         public bool RequiresPw { get; set; }
         public String Map { get; set; }
         public ServerSettings ServerSettings { get; set; }
